# storage/engine: optimize iteration over inline values