@RaduBerinde 2016-04-26
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Reviewed 1 of 1 files at r1.
Review status: all files reviewed at latest revision, 1 unresolved discussion, some commit checks pending.

---

*[storage/engine/mvcc.go, line 1643 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6320#-KGI59hvtj9vCcVI5ReF:-KGI59hvtj9vCcVI5ReG:983190472) ([raw file](https://github.com/cockroachdb/cockroach/blob/d82176d2b0959c113686abaec9293a8423f526bc/storage/engine/mvcc.go#L1643)):*
"versioned value"?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6320#-:-KGI5COz5t6fN-yPNopk:1417088721)*
<!-- Sent from Reviewable.io -->