@nvanbenschoten 2016-04-26
LGTM, but @tamird should review this as well because this seems to be exactly what we were looking for yesterday.