@petermattis 2016-04-26
Review status: all files reviewed at latest revision, 1 unresolved discussion, some commit checks pending.

---

*[storage/engine/mvcc.go, line 1643 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6320#-KGI59hvtj9vCcVI5ReF:-KGI5N5vItRPKp8DNfp4:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/d82176d2b0959c113686abaec9293a8423f526bc/storage/engine/mvcc.go#L1643)):*
Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6320)*
<!-- Sent from Reviewable.io -->