@petermattis 2016-04-27
Review status: all files reviewed at latest revision, 2 unresolved discussions, all commit checks successful.

---

*[storage/engine/mvcc.go, line 1638 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6320#-KGI4kegs7CuqodFxsOY-r2-1638:-KGNdVQaCNTWQMvZ1fHd:-2132107383) ([raw file](https://github.com/cockroachdb/cockroach/blob/7e6d02dd40c2613e250f3cb07279a80518b68dd1/storage/engine/mvcc.go#L1638)):*
What @tamird said.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6320)*
<!-- Sent from Reviewable.io -->